+++ ding/model/template/vac.py
@@ -5,8 +5,7 @@ import torch.nn as nn
 
 from ding.utils import SequenceType, squeeze, MODEL_REGISTRY
 from ..common import ReparameterizationHead, RegressionHead, DiscreteHead, MultiHead, \
-    FCEncoder, ConvEncoder
-from ding.torch_utils import ImpalaEncoder
+    FCEncoder, ConvEncoder, ImpalaCNN
 
 
 @MODEL_REGISTRY.register('vac')
@@ -36,7 +35,7 @@ class VAC(nn.Module):
         fixed_sigma_value: Optional[int] = 0.3,
         bound_type: Optional[str] = None,
         encoder: Optional[torch.nn.Module] = None,
-        use_impala_cnn_encoder: bool = False,
+        impala_cnn_encoder: bool = False,
     ) -> None:
         r"""
         Overview:
@@ -63,7 +62,7 @@ class VAC(nn.Module):
         obs_shape: int = squeeze(obs_shape)
         action_shape = squeeze(action_shape)
         self.obs_shape, self.action_shape = obs_shape, action_shape
-        self.use_impala_cnn_encoder = use_impala_cnn_encoder
+        self.impala_cnn_encoder = impala_cnn_encoder
         # Encoder Type
         if encoder is None:
             if isinstance(obs_shape, int) or len(obs_shape) == 1:
@@ -76,20 +75,26 @@ class VAC(nn.Module):
                     format(obs_shape)
                 )
             self.share_encoder = share_encoder
-            if self.share_encoder and self.use_impala_cnn_encoder:
-                encoder_cls = ImpalaEncoder
-                self.encoder = encoder_cls(obs_shape)
-            elif self.share_encoder:
-                self.encoder = encoder_cls(
-                    obs_shape, encoder_hidden_size_list, activation=activation, norm_type=norm_type
-                )
+            if self.share_encoder:
+                if self.impala_cnn_encoder:
+                    encoder_cls = ImpalaEncoder
+                    self.encoder = encoder_cls(obs_shape)
+                else:
+                    self.encoder = encoder_cls(
+                        obs_shape, encoder_hidden_size_list, activation=activation, norm_type=norm_type
+                    )
             else:
-                self.actor_encoder = encoder_cls(
-                    obs_shape, encoder_hidden_size_list, activation=activation, norm_type=norm_type
-                )
-                self.critic_encoder = encoder_cls(
-                    obs_shape, encoder_hidden_size_list, activation=activation, norm_type=norm_type
-                )
+                if self.impala_cnn_encoder:
+                    encoder_cls = ImpalaEncoder
+                    self.actor_encoder = encoder_cls(obs_shape)
+                    self.critic_encoder = encoder_cls(obs_shape)
+                else:
+                    self.actor_encoder = encoder_cls(
+                        obs_shape, encoder_hidden_size_list, activation=activation, norm_type=norm_type
+                    )
+                    self.critic_encoder = encoder_cls(
+                        obs_shape, encoder_hidden_size_list, activation=activation, norm_type=norm_type
+                    )
         else:
             assert share_encoder
             self.share_encoder = share_encoder
